7_Agent_Architecture/7.7-SupervisorAgent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
asana>=3.2.0
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.55.0
anyio==4.9.0
argcomplete==3.6.2
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.3.0
backoff==2.2.1
blinker==1.9.0
blockbuster==1.5.25
boto3==1.39.0
botocore==1.39.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
cohere==5.15.0
colorama==0.4.6
cryptography==44.0.3
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
eval_type_backport==0.2.2
fasta2a==0.3.5
fastapi==0.115.14
fastavro==1.11.1
filelock==3.18.0
forbiddenfruit==0.1.4
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.23.0
googleapis-common-protos==1.70.0
gotrue==2.12.3
greenlet==3.2.3
griffe==1.7.3
groq==0.29.0
grpcio==1.73.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jsonschema_rs==0.29.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langfuse==3.2.1
langgraph==0.5.3
langgraph-api==0.2.96
langgraph-checkpoint==2.1.0
langgraph-cli==0.3.5
langgraph-prebuilt==0.5.2
langgraph-runtime-inmem==0.6.0
langgraph-sdk==0.1.72
langsmith==0.4.4
logfire-api==3.21.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.10.1
mdurl==0.1.2
mistralai==1.9.1
multidict==6.6.3
mypy_extensions==1.1.0
narwhals==1.47.1
nest-asyncio==1.6.0
numpy==2.3.1
oauthlib==3.3.1
openai==1.93.0
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.1
pillow==11.3.0
pluggy==1.6.0
postgrest==1.1.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-ai==0.3.5
pydantic-ai-slim==0.3.5
pydantic-evals==0.3.5
pydantic-graph==0.3.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.5.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruff==0.12.1
s3transfer==0.13.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.1.3
starlette==0.46.2
storage3==0.12.0
streamlit==1.47.0
StrEnum==0.4.15
structlog==25.4.0
supabase==2.16.0
supafunc==0.10.1
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
truststore==0.10.1
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
